Remove unnecessary scrollbars from table
When using the developer tools to remove the `width`, `max-width` and `overflow-y` from the `.dash-spreadsheet-container.dash-spreadsheet` element it removed the unnecessary horizontal scrollbar as well as the duplicated vertical scrollbar on the right of the table. These doesn't scroll anything for me, but sometimes caused the table body and header to misalign. Worked for me in Chrome and Firefox on Desktop and didn't find any issues caused by this change.

Haven't tested other browsers or mobile, though and only guessed that this is the right place in the code to fix this, as I haven't executed it myself.